+++ lib/ch/row_binary.ex
@@ -94,6 +94,7 @@ defmodule Ch.RowBinary do
               :string,
               :binary,
               :json,
+              :dynamic,
               :boolean,
               :uuid,
               :date,
@@ -417,6 +418,19 @@ defmodule Ch.RowBinary do
   def encode(:polygon, rings), do: encode({:array, :ring}, rings)
   def encode(:multipolygon, polygons), do: encode({:array, :polygon}, polygons)
 
+  # TODO
+  def encode(:dynamic, value) do
+    case value do
+      _ when is_binary(value) -> [0x15 | encode(:string, value)]
+      _ when is_integer(value) and value >= 0 -> [0x04 | encode(:u64, value)]
+      _ when is_integer(value) -> [0x0A | encode(:i64, value)]
+      _ when is_float(value) -> [0x0E | encode(:f64, value)]
+      %Date{} -> [0x0F | encode(:date, value)]
+      %NaiveDateTime{} -> [0x11 | encode(:datetime, value)]
+      [] -> [0x1E, 0x00]
+    end
+  end
+
   # TODO enum8 and enum16 nil
   for size <- [8, 16] do
     enum_t = :"enum#{size}"
@@ -607,6 +621,7 @@ defmodule Ch.RowBinary do
               :string,
               :binary,
               :json,
+              :dynamic,
               :boolean,
               :uuid,
               :date,
@@ -862,6 +877,255 @@ defmodule Ch.RowBinary do
 
   defp map_types(0, _key_type, _value_types), do: []
 
+  # https://clickhouse.com/docs/sql-reference/data-types/data-types-binary-encoding
+  dynamic_types = [
+    nothing: 0x00,
+    u8: 0x01,
+    u16: 0x02,
+    u32: 0x03,
+    u64: 0x04,
+    u128: 0x05,
+    u256: 0x06,
+    i8: 0x07,
+    i16: 0x08,
+    i32: 0x09,
+    i64: 0x0A,
+    i128: 0x0B,
+    i256: 0x0C,
+    f32: 0x0D,
+    f64: 0x0E,
+    date: 0x0F,
+    date32: 0x10,
+    string: 0x15,
+    uuid: 0x1D,
+    ipv4: 0x28,
+    ipv6: 0x29,
+    boolean: 0x2D
+  ]
+
+  # TODO compile inline?
+
+  for {type, code} <- dynamic_types do
+    defp decode_dynamic(
+           <<unquote(code), rest::bytes>>,
+           dynamic,
+           types_rest,
+           row,
+           rows,
+           types
+         ) do
+      decode_dynamic_continue(rest, [unquote(type) | dynamic], types_rest, row, rows, types)
+    end
+  end
+
+  # DateTime 0x11
+  defp decode_dynamic(<<0x11, rest::bytes>>, dynamic, types_rest, row, rows, types) do
+    decode_dynamic_continue(
+      rest,
+      [{:datetime, nil} | dynamic],
+      types_rest,
+      row,
+      rows,
+      types
+    )
+  end
+
+  # DateTime(time_zone) 0x12 <var_uint_time_zone_name_size><time_zone_name_data>
+  for {pattern, size} <- varints do
+    defp decode_dynamic(
+           <<0x12, unquote(pattern), tz::size(unquote(size))-bytes, rest::bytes>>,
+           dynamic,
+           types_rest,
+           row,
+           rows,
+           types
+         ) do
+      decode_dynamic_continue(
+        rest,
+        [{:datetime, tz} | dynamic],
+        types_rest,
+        row,
+        rows,
+        types
+      )
+    end
+  end
+
+  # DateTime64(P) 0x13 <uint8_precision>
+  defp decode_dynamic(
+         <<0x13, precision, rest::bytes>>,
+         dynamic,
+         types_rest,
+         row,
+         rows,
+         types
+       ) do
+    decode_dynamic_continue(
+      rest,
+      [decoding_type({:datetime64, precision}) | dynamic],
+      types_rest,
+      row,
+      rows,
+      types
+    )
+  end
+
+  # DateTime64(P, time_zone) 0x14 <uint8_precision><var_uint_time_zone_name_size><time_zone_name_data>
+  for {pattern, size} <- varints do
+    defp decode_dynamic(
+           <<0x14, precision, unquote(pattern), tz::size(unquote(size))-bytes, rest::bytes>>,
+           dynamic,
+           types_rest,
+           row,
+           rows,
+           types
+         ) do
+      decode_dynamic_continue(
+        rest,
+        [decoding_type({:datetime64, precision, tz}) | dynamic],
+        types_rest,
+        row,
+        rows,
+        types
+      )
+    end
+  end
+
+  # FixedString(N) 0x16 <var_uint_size>
+  for {pattern, size} <- varints do
+    defp decode_dynamic(
+           <<0x16, unquote(pattern), rest::bytes>>,
+           dynamic,
+           types_rest,
+           row,
+           rows,
+           types
+         ) do
+      decode_dynamic_continue(
+        rest,
+        [{:fixed_string, unquote(size)} | dynamic],
+        types_rest,
+        row,
+        rows,
+        types
+      )
+    end
+  end
+
+  # Decimal32(P, S) 0x19 <uint8_precision><uint8_scale>
+  # Decimal64(P, S) 0x1A <uint8_precision><uint8_scale>
+  # Decimal128(P, S) 0x1B <uint8_precision><uint8_scale>
+  # Decimal256(P, S) 0x1C <uint8_precision><uint8_scale>
+  for {code, size} <- [
+        {0x19, 32},
+        {0x1A, 64},
+        {0x1B, 128},
+        {0x1C, 256}
+      ] do
+    defp decode_dynamic(
+           <<unquote(code), _precision, scale, rest::bytes>>,
+           dynamic,
+           types_rest,
+           row,
+           rows,
+           types
+         ) do
+      decode_dynamic_continue(
+        rest,
+        [{:decimal, unquote(size), scale} | dynamic],
+        types_rest,
+        row,
+        rows,
+        types
+      )
+    end
+  end
+
+  # Array(T) 0x1E <nested_type_encoding>
+  defp decode_dynamic(<<0x1E, rest::bytes>>, dynamic, types_rest, row, rows, types) do
+    decode_dynamic_continue(rest, [:array | dynamic], types_rest, row, rows, types)
+  end
+
+  # Nullable(T)	0x23 <nested_type_encoding>
+  defp decode_dynamic(<<0x23, rest::bytes>>, dynamic, types_rest, row, rows, types) do
+    decode_dynamic_continue(rest, [:nullable | dynamic], types_rest, row, rows, types)
+  end
+
+  # LowCardinality(T) 0x26 <nested_type_encoding>
+  defp decode_dynamic(<<0x26, rest::bytes>>, dynamic, types_rest, row, rows, types) do
+    decode_dynamic_continue(rest, [:low_cardinality | dynamic], types_rest, row, rows, types)
+  end
+
+  # TODO
+  # Enum8	0x17 <var_uint_number_of_elements><var_uint_name_size_1><name_data_1><int8_value_1>...<var_uint_name_size_N><name_data_N><int8_value_N>
+  # Enum16	0x18 <var_uint_number_of_elements><var_uint_name_size_1><name_data_1><int16_little_endian_value_1>...><var_uint_name_size_N><name_data_N><int16_little_endian_value_N>
+  # Tuple(T1, ..., TN)	0x1F <var_uint_number_of_elements><nested_type_encoding_1>...<nested_type_encoding_N>
+  # Tuple(name1 T1, ..., nameN TN)	0x20 <var_uint_number_of_elements><var_uint_name_size_1><name_data_1><nested_type_encoding_1>...<var_uint_name_size_N><name_data_N><nested_type_encoding_N>
+  # Set	0x21
+  # Interval	0x22 <interval_kind> (see interval kind binary encoding)
+  # Function	0x24<var_uint_number_of_arguments><argument_type_encoding_1>...<argument_type_encoding_N><return_type_encoding>
+  # AggregateFunction(function_name(param_1, ..., param_N), arg_T1, ..., arg_TN)	0x25<var_uint_version><var_uint_function_name_size><function_name_data><var_uint_number_of_parameters><param_1>...<param_N><var_uint_number_of_arguments><argument_type_encoding_1>...<argument_type_encoding_N> (see aggregate function parameter binary encoding)
+  # Map(K, V)	0x27<key_type_encoding><value_type_encoding>
+  # Variant(T1, ..., TN)	0x2A<var_uint_number_of_variants><variant_type_encoding_1>...<variant_type_encoding_N>
+  # Dynamic(max_types=N)	0x2B<uint8_max_types>
+  # Custom type (Ring, Polygon, etc)	0x2C<var_uint_type_name_size><type_name_data>
+  # SimpleAggregateFunction(function_name(param_1, ..., param_N), arg_T1, ..., arg_TN)	0x2E<var_uint_function_name_size><function_name_data><var_uint_number_of_parameters><param_1>...<param_N><var_uint_number_of_arguments><argument_type_encoding_1>...<argument_type_encoding_N> (see aggregate function parameter binary encoding)
+  # Nested(name1 T1, ..., nameN TN)	0x2F<var_uint_number_of_elements><var_uint_name_size_1><name_data_1><nested_type_encoding_1>...<var_uint_name_size_N><name_data_N><nested_type_encoding_N>
+  # JSON(max_dynamic_paths=N, max_dynamic_types=M, path Type, SKIP skip_path, SKIP REGEXP skip_path_regexp)	0x30<uint8_serialization_version><var_int_max_dynamic_paths><uint8_max_dynamic_types><var_uint_number_of_typed_paths><var_uint_path_name_size_1><path_name_data_1><encoded_type_1>...<var_uint_number_of_skip_paths><var_uint_skip_path_size_1><skip_path_data_1>...<var_uint_number_of_skip_path_regexps><var_uint_skip_path_regexp_size_1><skip_path_data_regexp_1>...
+
+  unsupported_dynamic_types = %{
+    "Enum8" => 0x17,
+    "Enum16" => 0x18,
+    "Tuple" => 0x1F,
+    "TupleWithNames" => 0x20,
+    "Set" => 0x21,
+    "Interval" => 0x22,
+    "Function" => 0x24,
+    "AggregateFunction" => 0x25,
+    "Map" => 0x27,
+    "Variant" => 0x2A,
+    "Dynamic" => 0x2B,
+    "CustomType" => 0x2C,
+    "SimpleAggregateFunction" => 0x2E,
+    "Nested" => 0x2F,
+    "JSON" => 0x30
+  }
+
+  for {type, code} <- unsupported_dynamic_types do
+    defp decode_dynamic(<<unquote(code), _::bytes>>, _dynamic, _types_rest, _row, _rows, _types) do
+      raise ArgumentError, "unsupported dynamic type #{unquote(type)}"
+    end
+  end
+
+  @compile inline: [decode_dynamic_continue: 6]
+
+  defp decode_dynamic_continue(<<rest::bytes>>, dynamic, types_rest, row, rows, types) do
+    continue? =
+      case dynamic do
+        [:array | _] -> true
+        [:nullable | _] -> true
+        [:low_cardinality | _] -> true
+        _ -> false
+      end
+
+    if continue? do
+      decode_dynamic(rest, dynamic, types_rest, row, rows, types)
+    else
+      type = build_dynamic_type(:lists.reverse(dynamic))
+      decode_rows([type | types_rest], rest, row, rows, types)
+    end
+  end
+
+  defp build_dynamic_type([type]), do: type
+
+  defp build_dynamic_type(type) do
+    case type do
+      [:array | rest] -> {:array, build_dynamic_type(rest)}
+      [:nullable | rest] -> {:nullable, build_dynamic_type(rest)}
+      [:low_cardinality | rest] -> build_dynamic_type(rest)
+    end
+  end
+
   defp decode_rows([type | types_rest], <<bin::bytes>>, row, rows, types) do
     case type do
       :u8 ->
@@ -941,6 +1205,9 @@ defmodule Ch.RowBinary do
         # i.e. assumes `settings: [output_format_binary_write_json_as_string: 1]`
         # TODO
         decode_string_json_decode_rows(bin, types_rest, row, rows, types)
+
+      :dynamic ->
+        decode_dynamic(bin, _dynamic = [], types_rest, row, rows, types)
 
       # TODO utf8?
       {:fixed_string, size} ->
@@ -1025,6 +1292,9 @@ defmodule Ch.RowBinary do
           <<1, bin::bytes>> -> decode_rows(types_rest, bin, [nil | row], rows, types)
           <<0, bin::bytes>> -> decode_rows([type | types_rest], bin, row, rows, types)
         end
+
+      :nothing ->
+        decode_rows(types_rest, bin, [nil | row], rows, types)
 
       {:array, type} ->
         decode_array_decode_rows(bin, type, types_rest, row, rows, types)